+++ main.py
@@ -1,0 +1,5 @@
+if __name__ == '__main__':
+    import sys
+
+    from package import app
+    sys.exit(app.run())
