+++ package/chatgptapi.py
@@ -1,357 +1,37 @@
-import ctypes
+
 import glob
 import os
-import platform
-import pprint
 import re
 import shutil
-import sqlite3
 import subprocess
 import sys
 import traceback
 import urllib.parse
 import webbrowser
 from datetime import datetime
-from functools import partial
 from io import StringIO
-from shutil import copyfile
 
 import openai
-import qdarktheme
 import tiktoken
 from gtts import gTTS
+from PySide6.QtCore import QRegularExpression, Qt, QThread, Signal
+from PySide6.QtGui import (QAction, QFontMetrics, QGuiApplication, QIcon,
+                           QStandardItem, QStandardItemModel, QTextDocument)
+from PySide6.QtPrintSupport import QPrintDialog, QPrinter
+from PySide6.QtWidgets import (QApplication, QCheckBox, QComboBox, QCompleter,
+                               QDialog, QDialogButtonBox, QFileDialog,
+                               QFormLayout, QHBoxLayout, QLabel, QLineEdit,
+                               QListView, QMainWindow, QMenu, QMessageBox,
+                               QPlainTextEdit, QProgressBar, QPushButton,
+                               QSplitter, QSystemTrayIcon, QVBoxLayout,
+                               QWidget)
 
 import config
-from configDefault import *
-from util.Worker import ChatGPTResponse, OpenAIImage
-
-# set working directory
-thisFile = os.path.realpath(__file__)
-wd = os.path.dirname(thisFile)
-if os.getcwd() != wd:
-    os.chdir(wd)
-if not os.path.isfile("config.py"):
-    open("config.py", "a", encoding="utf-8").close()
-try:
-    from pocketsphinx import LiveSpeech, get_model_path
-    isPocketsphinxInstalled = True
-except:
-    isPocketsphinxInstalled = False
-
-if config.qtLibrary == "pyside6":
-    from PySide6.QtCore import QRegularExpression, Qt, QThread, Signal
-    from PySide6.QtGui import (QAction, QFontMetrics, QGuiApplication, QIcon,
-                               QStandardItem, QStandardItemModel,
-                               QTextDocument)
-    from PySide6.QtPrintSupport import QPrintDialog, QPrinter
-    from PySide6.QtWidgets import (QApplication, QCheckBox, QComboBox,
-                                   QCompleter, QDialog, QDialogButtonBox,
-                                   QFileDialog, QFormLayout, QHBoxLayout,
-                                   QLabel, QLineEdit, QListView, QMainWindow,
-                                   QMenu, QMessageBox, QPlainTextEdit,
-                                   QProgressBar, QPushButton, QSplitter,
-                                   QSystemTrayIcon, QVBoxLayout, QWidget)
-else:
-    from qtpy.QtCore import QRegularExpression, Qt, QThread, Signal
-    from qtpy.QtGui import (QFontMetrics, QGuiApplication, QIcon,
-                            QStandardItem, QStandardItemModel, QTextDocument)
-    from qtpy.QtPrintSupport import QPrintDialog, QPrinter
-    from qtpy.QtWidgets import (QAction, QApplication, QCheckBox, QComboBox,
-                                QCompleter, QDialog, QDialogButtonBox,
-                                QFileDialog, QFormLayout, QHBoxLayout, QLabel,
-                                QLineEdit, QListView, QMainWindow, QMenu,
-                                QMessageBox, QPlainTextEdit, QProgressBar,
-                                QPushButton, QSplitter, QSystemTrayIcon,
-                                QVBoxLayout, QWidget)
-
-
-class SpeechRecognitionThread(QThread):
-    phrase_recognized = Signal(str)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.is_running = False
-
-    def run(self):
-        self.is_running = True
-        if config.pocketsphinxModelPath:
-            # download English dictionary at: http://www.speech.cs.cmu.edu/cgi-bin/cmudict
-            # download voice models at https://sourceforge.net/projects/cmusphinx/files/Acoustic%20and%20Language%20Models/
-            speech = LiveSpeech(
-                # sampling_rate=16000,  # optional
-                hmm=get_model_path(config.pocketsphinxModelPath),
-                lm=get_model_path(config.pocketsphinxModelPathBin),
-                dic=get_model_path(config.pocketsphinxModelPathDict),
-            )
-        else:
-            speech = LiveSpeech()
-
-        for phrase in speech:
-            if not self.is_running:
-                break
-            recognized_text = str(phrase)
-            self.phrase_recognized.emit(recognized_text)
-
-    def stop(self):
-        self.is_running = False
-
-
-class ApiDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle(config.thisTranslation["settings"])
-
-        self.apiKeyEdit = QLineEdit(config.openaiApiKey)
-        self.apiKeyEdit.setEchoMode(QLineEdit.Password)
-        self.orgEdit = QLineEdit(config.openaiApiOrganization)
-        self.orgEdit.setEchoMode(QLineEdit.Password)
-        self.apiModelBox = QComboBox()
-        initialIndex = 0
-        index = 0
-        for key in ("gpt-3.5-turbo", "gpt-3.5-turbo-16k", "gpt-4", "gpt-4-32k"):
-            self.apiModelBox.addItem(key)
-            if key == config.chatGPTApiModel:
-                initialIndex = index
-            index += 1
-        self.apiModelBox.setCurrentIndex(initialIndex)
-        self.functionCallingBox = QComboBox()
-        initialIndex = 0
-        index = 0
-        for key in ("auto", "none"):
-            self.functionCallingBox.addItem(key)
-            if key == config.chatGPTApiFunctionCall:
-                initialIndex = index
-            index += 1
-        self.functionCallingBox.setCurrentIndex(initialIndex)
-        self.loadingInternetSearchesBox = QComboBox()
-        initialIndex = 0
-        index = 0
-        for key in ("always", "auto", "none"):
-            self.loadingInternetSearchesBox.addItem(key)
-            if key == config.loadingInternetSearches:
-                initialIndex = index
-            index += 1
-        self.loadingInternetSearchesBox.setCurrentIndex(initialIndex)
-        self.maxTokenEdit = QLineEdit(str(config.chatGPTApiMaxTokens))
-        self.maxTokenEdit.setToolTip(
-            "The maximum number of tokens to generate in the completion.\nThe token count of your prompt plus max_tokens cannot exceed the model's context length. Most models have a context length of 2048 tokens (except for the newest models, which support 4096).")
-        self.maxInternetSearchResults = QLineEdit(
-            str(config.maximumInternetSearchResults))
-        self.maxInternetSearchResults.setToolTip(
-            "The maximum number of internet search response to be included.")
-        # self.includeInternetSearches = QCheckBox(config.thisTranslation["include"])
-        # self.includeInternetSearches.setToolTip("Include latest internet search results")
-        # self.includeInternetSearches.setCheckState(Qt.Checked if config.includeDuckDuckGoSearchResults else Qt.Unchecked)
-        # self.includeDuckDuckGoSearchResults = config.includeDuckDuckGoSearchResults
-        self.autoScrollingCheckBox = QCheckBox(
-            config.thisTranslation["enable"])
-        self.autoScrollingCheckBox.setToolTip(
-            "Auto-scroll display as responses are received")
-        self.autoScrollingCheckBox.setCheckState(
-            Qt.Checked if config.chatGPTApiAutoScrolling else Qt.Unchecked)
-        self.chatGPTApiAutoScrolling = config.chatGPTApiAutoScrolling
-        self.chatAfterFunctionCalledCheckBox = QCheckBox(
-            config.thisTranslation["enable"])
-        self.chatAfterFunctionCalledCheckBox.setToolTip(
-            "Automatically generate next chat response after a function is called")
-        self.chatAfterFunctionCalledCheckBox.setCheckState(
-            Qt.Checked if config.chatAfterFunctionCalled else Qt.Unchecked)
-        self.chatAfterFunctionCalled = config.chatAfterFunctionCalled
-        self.runPythonScriptGloballyCheckBox = QCheckBox(
-            config.thisTranslation["enable"])
-        self.runPythonScriptGloballyCheckBox.setToolTip(
-            "Run user python script in global scope")
-        self.runPythonScriptGloballyCheckBox.setCheckState(
-            Qt.Checked if config.runPythonScriptGlobally else Qt.Unchecked)
-        self.runPythonScriptGlobally = config.runPythonScriptGlobally
-        self.contextEdit = QLineEdit(config.chatGPTApiContext)
-        firstInputOnly = config.thisTranslation["firstInputOnly"]
-        allInputs = config.thisTranslation["allInputs"]
-        self.applyContextIn = QComboBox()
-        self.applyContextIn.addItems([firstInputOnly, allInputs])
-        self.applyContextIn.setCurrentIndex(
-            1 if config.chatGPTApiContextInAllInputs else 0)
-        self.predefinedContextBox = QComboBox()
-        initialIndex = 0
-        index = 0
-        for key, value in config.predefinedContexts.items():
-            self.predefinedContextBox.addItem(key)
-            self.predefinedContextBox.setItemData(
-                self.predefinedContextBox.count()-1, value, role=Qt.ToolTipRole)
-            if key == config.chatGPTApiPredefinedContext:
-                initialIndex = index
-            index += 1
-        self.predefinedContextBox.currentIndexChanged.connect(
-            self.predefinedContextBoxChanged)
-        self.predefinedContextBox.setCurrentIndex(initialIndex)
-        # set availability of self.contextEdit in case there is no index changed
-        self.contextEdit.setDisabled(
-            True) if not initialIndex == 1 else self.contextEdit.setEnabled(True)
-        buttonBox = QDialogButtonBox(
-            QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        buttonBox.accepted.connect(self.accept)
-        buttonBox.rejected.connect(self.reject)
-
-        layout = QFormLayout()
-        # https://platform.openai.com/account/api-keys
-        chatAfterFunctionCalled = config.thisTranslation["chatAfterFunctionCalled"]
-        runPythonScriptGlobally = config.thisTranslation["runPythonScriptGlobally"]
-        autoScroll = config.thisTranslation["autoScroll"]
-        predefinedContext = config.thisTranslation["predefinedContext"]
-        context = config.thisTranslation["chatContext"]
-        applyContext = config.thisTranslation["applyContext"]
-        latestOnlineSearchResults = config.thisTranslation["latestOnlineSearchResults"]
-        maximumOnlineSearchResults = config.thisTranslation["maximumOnlineSearchResults"]
-        required = config.thisTranslation["required"]
-        optional = config.thisTranslation["optional"]
-        layout.addRow(f"OpenAI API Key [{required}]:", self.apiKeyEdit)
-        layout.addRow(f"Organization ID [{optional}]:", self.orgEdit)
-        layout.addRow(f"API Model [{required}]:", self.apiModelBox)
-        layout.addRow(f"Max Token [{required}]:", self.maxTokenEdit)
-        layout.addRow(
-            f"Function Calling [{optional}]:", self.functionCallingBox)
-        layout.addRow(
-            f"{chatAfterFunctionCalled} [{optional}]:", self.chatAfterFunctionCalledCheckBox)
-        layout.addRow(
-            f"{predefinedContext} [{optional}]:", self.predefinedContextBox)
-        layout.addRow(f"{context} [{optional}]:", self.contextEdit)
-        layout.addRow(f"{applyContext} [{optional}]:", self.applyContextIn)
-        layout.addRow(
-            f"{latestOnlineSearchResults} [{optional}]:", self.loadingInternetSearchesBox)
-        layout.addRow(
-            f"{maximumOnlineSearchResults} [{optional}]:", self.maxInternetSearchResults)
-        layout.addRow(f"{autoScroll} [{optional}]:",
-                      self.autoScrollingCheckBox)
-        layout.addRow(
-            f"{runPythonScriptGlobally} [{optional}]:", self.runPythonScriptGloballyCheckBox)
-        layout.addWidget(buttonBox)
-        self.autoScrollingCheckBox.stateChanged.connect(
-            self.toggleAutoScrollingCheckBox)
-        self.chatAfterFunctionCalledCheckBox.stateChanged.connect(
-            self.toggleChatAfterFunctionCalled)
-        self.runPythonScriptGloballyCheckBox.stateChanged.connect(
-            self.toggleRunPythonScriptGlobally)
-        self.functionCallingBox.currentIndexChanged.connect(
-            self.functionCallingBoxChanged)
-        self.loadingInternetSearchesBox.currentIndexChanged.connect(
-            self.loadingInternetSearchesBoxChanged)
-
-        self.setLayout(layout)
-
-    def api_key(self):
-        return self.apiKeyEdit.text().strip()
-
-    def org(self):
-        return self.orgEdit.text().strip()
-
-    def context(self):
-        return self.contextEdit.text().strip()
-
-    def contextInAllInputs(self):
-        return True if self.applyContextIn.currentIndex() == 1 else False
-
-    def predefinedContextBoxChanged(self, index):
-        self.contextEdit.setDisabled(
-            True) if not index == 1 else self.contextEdit.setEnabled(True)
-
-    def predefinedContext(self):
-        return self.predefinedContextBox.currentText()
-        # return self.predefinedContextBox.currentData(Qt.ToolTipRole)
-
-    def apiModel(self):
-        # return "gpt-3.5-turbo"
-        return self.apiModelBox.currentText()
-
-    def functionCalling(self):
-        return self.functionCallingBox.currentText()
-
-    def max_token(self):
-        return self.maxTokenEdit.text().strip()
-
-    def enable_auto_scrolling(self):
-        return self.chatGPTApiAutoScrolling
-
-    def toggleAutoScrollingCheckBox(self, state):
-        self.chatGPTApiAutoScrolling = True if state else False
-
-    def enable_chatAfterFunctionCalled(self):
-        return self.chatAfterFunctionCalled
-
-    def toggleChatAfterFunctionCalled(self, state):
-        self.chatAfterFunctionCalled = True if state else False
-
-    def enable_runPythonScriptGlobally(self):
-        return self.runPythonScriptGlobally
-
-    def toggleRunPythonScriptGlobally(self, state):
-        self.runPythonScriptGlobally = True if state else False
-
-    def functionCallingBoxChanged(self):
-        if self.functionCallingBox.currentText() == "none" and self.loadingInternetSearchesBox.currentText() == "auto":
-            self.loadingInternetSearchesBox.setCurrentText("none")
-
-    def loadingInternetSearches(self):
-        return self.loadingInternetSearchesBox.currentText()
-
-    def loadingInternetSearchesBoxChanged(self, _):
-        if self.loadingInternetSearchesBox.currentText() == "auto":
-            self.functionCallingBox.setCurrentText("auto")
-
-    def max_token(self):
-        return self.maxTokenEdit.text().strip()
-
-    def max_internet_search_results(self):
-        return self.maxInternetSearchResults.text().strip()
-
-
-class Database:
-    def __init__(self, filePath=""):
-        def regexp(expr, item):
-            reg = re.compile(expr, flags=re.IGNORECASE)
-            return reg.search(item) is not None
-        defaultFilePath = config.chatGPTApiLastChatDatabase if config.chatGPTApiLastChatDatabase and os.path.isfile(
-            config.chatGPTApiLastChatDatabase) else os.path.join(wd, "chats", "default.chat")
-        self.filePath = filePath if filePath else defaultFilePath
-        self.connection = sqlite3.connect(self.filePath)
-        self.connection.create_function("REGEXP", 2, regexp)
-        self.cursor = self.connection.cursor()
-        self.cursor.execute(
-            'CREATE TABLE IF NOT EXISTS data (id TEXT PRIMARY KEY, title TEXT, content TEXT)')
-        self.connection.commit()
-
-    def insert(self, id, title, content):
-        self.cursor.execute('SELECT * FROM data WHERE id = ?', (id,))
-        existing_data = self.cursor.fetchone()
-        if existing_data:
-            if existing_data[1] == title and existing_data[2] == content:
-                return
-            else:
-                self.cursor.execute(
-                    'UPDATE data SET title = ?, content = ? WHERE id = ?', (title, content, id))
-                self.connection.commit()
-        else:
-            self.cursor.execute(
-                'INSERT INTO data (id, title, content) VALUES (?, ?, ?)', (id, title, content))
-            self.connection.commit()
-
-    def search(self, title, content):
-        if config.regexpSearchEnabled:
-            # with regular expression
-            self.cursor.execute(
-                'SELECT * FROM data WHERE title REGEXP ? AND content REGEXP ?', (title, content))
-        else:
-            # without regular expression
-            self.cursor.execute('SELECT * FROM data WHERE title LIKE ? AND content LIKE ?',
-                                ('%{}%'.format(title), '%{}%'.format(content)))
-        return self.cursor.fetchall()
-
-    def delete(self, id):
-        self.cursor.execute('DELETE FROM data WHERE id = ?', (id,))
-        self.connection.commit()
-
-    def clear(self):
-        self.cursor.execute('DELETE FROM data')
-        self.connection.commit()
+from package import isPocketsphinxInstalled
+from package.database import Database
+from package.dialogs.apidialog import ApiDialog
+from package.speech import SpeechRecognitionThread
+from package.util.Worker import ChatGPTResponse, OpenAIImage
 
 
 class ChatGPTAPI(QWidget):
@@ -1255,410 +935,3 @@ Follow the following steps:
 
     def closeMediaPlayer(self):
         pass
-
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        super().__init__()
-        self.initUI()
-
-    def reloadMenubar(self):
-        self.menuBar().clear()
-        self.createMenubar()
-
-    def createMenubar(self):
-        # Create a menu bar
-        menubar = self.menuBar()
-
-        # Create a File menu and add it to the menu bar
-        file_menu = menubar.addMenu(config.thisTranslation["chat"])
-
-        new_action = QAction(config.thisTranslation["openDatabase"], self)
-        new_action.setShortcut("Ctrl+Shift+O")
-        new_action.triggered.connect(self.chatGPT.openDatabase)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["newDatabase"], self)
-        new_action.setShortcut("Ctrl+Shift+N")
-        new_action.triggered.connect(self.chatGPT.newDatabase)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["saveDatabaseAs"], self)
-        new_action.setShortcut("Ctrl+Shift+S")
-        new_action.triggered.connect(
-            lambda: self.chatGPT.newDatabase(copyExistingDatabase=True))
-        file_menu.addAction(new_action)
-
-        file_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["fileManager"], self)
-        new_action.triggered.connect(self.openDatabaseDirectory)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["pluginDirectory"], self)
-        new_action.triggered.connect(self.openPluginsDirectory)
-        file_menu.addAction(new_action)
-
-        file_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["newChat"], self)
-        new_action.setShortcut("Ctrl+N")
-        new_action.triggered.connect(self.chatGPT.newData)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["saveChat"], self)
-        new_action.setShortcut("Ctrl+S")
-        new_action.triggered.connect(self.chatGPT.saveData)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["exportChat"], self)
-        new_action.triggered.connect(self.chatGPT.exportData)
-        file_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["printChat"], self)
-        new_action.setShortcut("Ctrl+P")
-        new_action.triggered.connect(self.chatGPT.printData)
-        file_menu.addAction(new_action)
-
-        file_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["readTextFile"], self)
-        new_action.triggered.connect(self.chatGPT.openTextFileDialog)
-        file_menu.addAction(new_action)
-
-        file_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["countPromptTokens"], self)
-        new_action.triggered.connect(self.chatGPT.num_tokens_from_messages)
-        file_menu.addAction(new_action)
-
-        file_menu.addSeparator()
-
-        # Create a Exit action and add it to the File menu
-        exit_action = QAction(config.thisTranslation["exit"], self)
-        exit_action.setShortcut("Ctrl+Q")
-        exit_action.setStatusTip(config.thisTranslation["exitTheApplication"])
-        exit_action.triggered.connect(QGuiApplication.instance().quit)
-        file_menu.addAction(exit_action)
-
-        # Create customise menu
-        customise_menu = menubar.addMenu(config.thisTranslation["customise"])
-
-        openSettings = QAction(config.thisTranslation["configure"], self)
-        openSettings.triggered.connect(self.chatGPT.showApiDialog)
-        customise_menu.addAction(openSettings)
-
-        customise_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["toggleDarkTheme"], self)
-        new_action.triggered.connect(self.toggleTheme)
-        customise_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["toggleSystemTray"], self)
-        new_action.triggered.connect(self.toggleSystemTray)
-        customise_menu.addAction(new_action)
-
-        new_action = QAction(
-            config.thisTranslation["toggleMultilineInput"], self)
-        new_action.setShortcut("Ctrl+L")
-        new_action.triggered.connect(self.chatGPT.multilineButtonClicked)
-        customise_menu.addAction(new_action)
-
-        new_action = QAction(config.thisTranslation["toggleRegexp"], self)
-        new_action.setShortcut("Ctrl+E")
-        new_action.triggered.connect(self.toggleRegexp)
-        customise_menu.addAction(new_action)
-
-        # Create predefined context menu
-        context_menu = menubar.addMenu(
-            config.thisTranslation["predefinedContext"])
-        for index, context in enumerate(config.predefinedContexts):
-            contextAction = QAction(context, self)
-            if index < 10:
-                contextAction.setShortcut(f"Ctrl+{index}")
-            contextAction.triggered.connect(
-                partial(self.chatGPT.bibleChatAction, context))
-            context_menu.addAction(contextAction)
-
-        # Create a plugin menu
-        plugin_menu = menubar.addMenu(config.thisTranslation["plugins"])
-
-        pluginFolder = os.path.join(os.getcwd(), "plugins")
-        for index, plugin in enumerate(self.fileNamesWithoutExtension(pluginFolder, "py")):
-            new_action = QAction(plugin, self)
-            new_action.setCheckable(True)
-            new_action.setChecked(
-                False if plugin in config.chatGPTPluginExcludeList else True)
-            new_action.triggered.connect(
-                partial(self.updateExcludePluginList, plugin))
-            plugin_menu.addAction(new_action)
-
-        # Create a text selection menu
-        text_selection_menu = menubar.addMenu(
-            config.thisTranslation["textSelection"])
-
-        new_action = QAction(config.thisTranslation["webBrowser"], self)
-        new_action.triggered.connect(self.chatGPT.webBrowse)
-        text_selection_menu.addAction(new_action)
-
-        new_action = QAction(
-            config.thisTranslation["runAsPythonCommand"], self)
-        new_action.triggered.connect(self.chatGPT.runPythonCommand)
-        text_selection_menu.addAction(new_action)
-
-        new_action = QAction(
-            config.thisTranslation["runAsSystemCommand"], self)
-        new_action.triggered.connect(self.chatGPT.runSystemCommand)
-        text_selection_menu.addAction(new_action)
-
-        # Create About menu
-        about_menu = menubar.addMenu(config.thisTranslation["about"])
-
-        openSettings = QAction(config.thisTranslation["repository"], self)
-        openSettings.triggered.connect(lambda: webbrowser.open(
-            "https://github.com/eliranwong/ChatGPT-GUI"))
-        about_menu.addAction(openSettings)
-
-        about_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["help"], self)
-        new_action.triggered.connect(lambda: webbrowser.open(
-            "https://github.com/eliranwong/ChatGPT-GUI/wiki"))
-        about_menu.addAction(new_action)
-
-        about_menu.addSeparator()
-
-        new_action = QAction(config.thisTranslation["donate"], self)
-        new_action.triggered.connect(lambda: webbrowser.open(
-            "https://www.paypal.com/paypalme/MarvelBible"))
-        about_menu.addAction(new_action)
-
-    def initUI(self):
-        # Set a central widget
-        self.chatGPT = ChatGPTAPI(self)
-        self.setCentralWidget(self.chatGPT)
-
-        # create menu bar
-        self.createMenubar()
-
-        # set initial window size
-        # self.setWindowTitle("ChatGPT-GUI")
-        self.resize(QGuiApplication.primaryScreen().availableSize() * 3 / 4)
-        self.show()
-
-    def updateExcludePluginList(self, plugin):
-        if plugin in config.chatGPTPluginExcludeList:
-            config.chatGPTPluginExcludeList.remove(plugin)
-        else:
-            config.chatGPTPluginExcludeList.append(plugin)
-        internetSeraches = "integrate google searches"
-        if internetSeraches in config.chatGPTPluginExcludeList and config.loadingInternetSearches == "auto":
-            config.loadingInternetSearches = "none"
-        elif not internetSeraches in config.chatGPTPluginExcludeList and config.loadingInternetSearches == "none":
-            config.loadingInternetSearches = "auto"
-            config.chatGPTApiFunctionCall = "auto"
-        # reload plugins
-        config.chatGPTApi.runPlugins()
-
-    def fileNamesWithoutExtension(self, dir, ext):
-        files = glob.glob(os.path.join(dir, "*.{0}".format(ext)))
-        return sorted([file[len(dir)+1:-(len(ext)+1)] for file in files if os.path.isfile(file)])
-
-    def getOpenCommand(self):
-        thisOS = platform.system()
-        if thisOS == "Windows":
-            openCommand = "start"
-        elif thisOS == "Darwin":
-            openCommand = "open"
-        elif thisOS == "Linux":
-            openCommand = "xdg-open"
-        return openCommand
-
-    def openDatabaseDirectory(self):
-        databaseDirectory = os.path.dirname(
-            os.path.abspath(config.chatGPTApiLastChatDatabase))
-        openCommand = self.getOpenCommand()
-        os.system(f"{openCommand} {databaseDirectory}")
-
-    def openPluginsDirectory(self):
-        openCommand = self.getOpenCommand()
-        os.system(f"{openCommand} plugins")
-
-    def toggleRegexp(self):
-        config.regexpSearchEnabled = not config.regexpSearchEnabled
-        self.chatGPT.updateSearchToolTips()
-        QMessageBox.information(
-            self, "ChatGPT-GUI", f"Regular expression for search and replace is {'enabled' if config.regexpSearchEnabled else 'disabled'}!")
-
-    def toggleSystemTray(self):
-        config.enableSystemTray = not config.enableSystemTray
-        QMessageBox.information(
-            self, "ChatGPT-GUI", "You need to restart this application to make the changes effective.")
-
-    def toggleTheme(self):
-        config.darkTheme = not config.darkTheme
-        qdarktheme.setup_theme() if config.darkTheme else qdarktheme.setup_theme("light")
-
-    # Work with system tray
-    def isWayland(self):
-        if platform.system() == "Linux" and not os.getenv('QT_QPA_PLATFORM') is None and os.getenv('QT_QPA_PLATFORM') == "wayland":
-            return True
-        else:
-            return False
-
-    def bringToForeground(self, window):
-        if window and not (window.isVisible() and window.isActiveWindow()):
-            window.raise_()
-            # Method activateWindow() does not work with qt.qpa.wayland
-            # platform.system() == "Linux" and not os.getenv('QT_QPA_PLATFORM') is None and os.getenv('QT_QPA_PLATFORM') == "wayland"
-            # The error message is received when QT_QPA_PLATFORM=wayland:
-            # qt.qpa.wayland: Wayland does not support QWindow::requestActivate()
-            # Therefore, we use hide and show methods instead with wayland.
-            if window.isVisible() and not window.isActiveWindow():
-                window.hide()
-            window.show()
-            if not self.isWayland():
-                window.activateWindow()
-
-
-if __name__ == '__main__':
-    def showMainWindow():
-        if not hasattr(config, "mainWindow") or config.mainWindow is None:
-            config.mainWindow = MainWindow()
-            qdarktheme.setup_theme() if config.darkTheme else qdarktheme.setup_theme("light")
-        else:
-            config.mainWindow.bringToForeground(config.mainWindow)
-
-    def aboutToQuit():
-        with open("config.py", "w", encoding="utf-8") as fileObj:
-            for name in dir(config):
-                excludeFromSavingList = (
-                    "mainWindow",  # main window object
-                    "chatGPTApi",  # GUI object
-                    "chatGPTTransformers",  # used with plugins; transform ChatGPT response message
-                    "predefinedContexts",  # used with plugins; pre-defined contexts
-                    "inputSuggestions",  # used with plugins; user input suggestions
-                    "integrate_google_searches_signature",
-                    "chatGPTApiFunctionSignatures",  # used with plugins; function calling
-                    "chatGPTApiAvailableFunctions",  # used with plugins; function calling
-                    # used with plugins; function calling when function name is 'python'
-                    "pythonFunctionResponse",
-                )
-                if not name.startswith("__") and not name in excludeFromSavingList:
-                    try:
-                        value = eval(f"config.{name}")
-                        fileObj.write("{0} = {1}\n".format(
-                            name, pprint.pformat(value)))
-                    except:
-                        pass
-
-    thisOS = platform.system()
-    appName = "ChatGPT-GUI"
-    # Windows icon
-    if thisOS == "Windows":
-        myappid = "chatgpt.gui"
-        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(myappid)
-        windowsIconPath = os.path.abspath(
-            os.path.join(sys.path[0], "icons", f"{appName}.ico"))
-        ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(
-            windowsIconPath)
-    # app
-    qdarktheme.enable_hi_dpi()
-    app = QApplication(sys.argv)
-    iconPath = os.path.abspath(os.path.join(
-        sys.path[0], "icons", f"{appName}.png"))
-    appIcon = QIcon(iconPath)
-    app.setWindowIcon(appIcon)
-    showMainWindow()
-    # connection
-    app.aboutToQuit.connect(aboutToQuit)
-
-    # Desktop shortcut
-    # on Windows
-    if thisOS == "Windows":
-        desktopPath = os.path.join(os.path.expanduser('~'), 'Desktop')
-        shortcutDir = desktopPath if os.path.isdir(desktopPath) else wd
-        shortcutBat1 = os.path.join(shortcutDir, f"{appName}.bat")
-        shortcutCommand1 = f'''powershell.exe -NoExit -Command "python '{thisFile}'"'''
-        # Create .bat for application shortcuts
-        if not os.path.exists(shortcutBat1):
-            try:
-                with open(shortcutBat1, "w") as fileObj:
-                    fileObj.write(shortcutCommand1)
-            except:
-                pass
-    # on macOS
-    elif thisOS == "Darwin":
-        shortcut_file = os.path.expanduser(f"~/Desktop/{appName}.command")
-        if not os.path.isfile(shortcut_file):
-            with open(shortcut_file, "w") as f:
-                f.write("#!/bin/bash\n")
-                f.write(f"cd {wd}\n")
-                f.write(f"{sys.executable} {thisFile} gui\n")
-            os.chmod(shortcut_file, 0o755)
-    # additional shortcuts on Linux
-    elif thisOS == "Linux":
-        def desktopFileContent():
-            iconPath = os.path.join(wd, "icons", "ChatGPT-GUI.png")
-            return """#!/usr/bin/env xdg-open
-
-[Desktop Entry]
-Version=1.0
-Type=Application
-Terminal=false
-Path={0}
-Exec={1} {2}
-Icon={3}
-Name=ChatGPT GUI
-""".format(wd, sys.executable, thisFile, iconPath)
-
-        ubaLinuxDesktopFile = os.path.join(wd, f"{appName}.desktop")
-        if not os.path.exists(ubaLinuxDesktopFile):
-            # Create .desktop shortcut
-            with open(ubaLinuxDesktopFile, "w") as fileObj:
-                fileObj.write(desktopFileContent())
-            try:
-                # Try to copy the newly created .desktop file to:
-                from pathlib import Path
-
-                # ~/.local/share/applications
-                userAppDir = os.path.join(
-                    str(Path.home()), ".local", "share", "applications")
-                userAppDirShortcut = os.path.join(
-                    userAppDir, f"{appName}.desktop")
-                if not os.path.exists(userAppDirShortcut):
-                    Path(userAppDir).mkdir(parents=True, exist_ok=True)
-                    copyfile(ubaLinuxDesktopFile, userAppDirShortcut)
-                # ~/Desktop
-                homeDir = os.environ["HOME"]
-                desktopPath = f"{homeDir}/Desktop"
-                desktopPathShortcut = os.path.join(
-                    desktopPath, f"{appName}.desktop")
-                if os.path.isfile(desktopPath) and not os.path.isfile(desktopPathShortcut):
-                    copyfile(ubaLinuxDesktopFile, desktopPathShortcut)
-            except:
-                pass
-
-    # system tray
-    if config.enableSystemTray:
-        app.setQuitOnLastWindowClosed(False)
-        # Set up tray icon
-        tray = QSystemTrayIcon()
-        tray.setIcon(appIcon)
-        tray.setToolTip("ChatGPT-GUI")
-        tray.setVisible(True)
-        # Import system tray menu
-        trayMenu = QMenu()
-        showMainWindowAction = QAction(config.thisTranslation["show"])
-        showMainWindowAction.triggered.connect(showMainWindow)
-        trayMenu.addAction(showMainWindowAction)
-        # Add a separator
-        trayMenu.addSeparator()
-        # Quit
-        quitAppAction = QAction(config.thisTranslation["exit"])
-        quitAppAction.triggered.connect(app.quit)
-        trayMenu.addAction(quitAppAction)
-        tray.setContextMenu(trayMenu)
-
-    # run the app
-    sys.exit(app.exec() if config.qtLibrary == "pyside6" else app.exec_())
