+++ package/speech.py
@@ -1,0 +1,37 @@
+from PySide6.QtCore import QThread, Signal
+
+from package import config, isPocketsphinxInstalled
+
+if isPocketsphinxInstalled:
+    from pocketsphinx import LiveSpeech, get_model_path
+
+
+class SpeechRecognitionThread(QThread):
+    phrase_recognized = Signal(str)
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.is_running = False
+
+    def run(self):
+        self.is_running = True
+        if config.pocketsphinxModelPath:
+            # download English dictionary at: http://www.speech.cs.cmu.edu/cgi-bin/cmudict
+            # download voice models at https://sourceforge.net/projects/cmusphinx/files/Acoustic%20and%20Language%20Models/
+            speech = LiveSpeech(
+                # sampling_rate=16000,  # optional
+                hmm=get_model_path(config.pocketsphinxModelPath),
+                lm=get_model_path(config.pocketsphinxModelPathBin),
+                dic=get_model_path(config.pocketsphinxModelPathDict),
+            )
+        else:
+            speech = LiveSpeech()
+
+        for phrase in speech:
+            if not self.is_running:
+                break
+            recognized_text = str(phrase)
+            self.phrase_recognized.emit(recognized_text)
+
+    def stop(self):
+        self.is_running = False
